% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agroindexGrid.R
\name{get_shape}
\alias{get_shape}
\title{Get shape/dimensions of a grid}
\usage{
get_shape(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[transformeR]{getShape}}}
}
\value{
A list or vector containing the dimensions of the grid (e.g., time,
lat, lon, member)
}
\description{
Wrapper function for \code{\link[transformeR]{getShape}}
to extract the dimensions (shape) of a climate grid object.
}
\seealso{
\code{\link[transformeR]{getShape}}
}
