% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agroindexGrid.R
\name{redim_tr}
\alias{redim_tr}
\title{Redimension a grid}
\usage{
redim_tr(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[transformeR]{redim}}}
}
\value{
A redimensioned grid object
}
\description{
Wrapper function for \code{\link[transformeR]{redim}}
to reshape or redimension a climate grid object by adding or removing dimensions.
}
\seealso{
\code{\link[transformeR]{redim}}
}
